// types/amp_gauges/eris_gov_amp_gauges_instantiate.d.ts
/* tslint:disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

/**
 * This structure describes the basic settings for creating a contract.
 */
export interface InstantiateMsg {
  /**
   * The vAMP token contract address
   */
  escrow_addr: string;
  /**
   * Hub contract address
   */
  hub_addr: string;
  /**
   * Contract owner
   */
  owner: string;
  /**
   * Max number of validators that can receive ASTRO emissions at the same time
   */
  validators_limit: number;
}